==> 1.training-neural-classifier.ipynb <==
# Training notebook for easy vs. standard German classifier 
# Authors: Hadi Asghari & Freya Hewett
# Version: 2023.02

# Note: due to copyright reasons we can only release a subset of our dataset, therefore if you re-run this 
# notebook, the results may differ

import pickle
import numpy as np
import pandas as pd
import tensorflow as tf
from tensorflow import keras
from tensorflow.keras import layers
from tensorflow.keras.layers import TextVectorization
from keras.callbacks import ModelCheckpoint

# Limit GPU memory use by TF
gpus = tf.config.experimental.list_physical_devices('GPU')
if gpus:
    print(gpus)
    tf.config.experimental.set_memory_growth(gpus[0], True)

# LOAD DATASETS  
train_pds = pd.read_csv("./data/dataset-train.csv", sep="\t")
print("train:", len(train_pds), "\n", train_pds.head(1))
eval_pds = pd.read_csv("./data/datset-test.csv", sep="\t")
print("validation:" , len(eval_pds), "\n", eval_pds.head(1))

# TURN DATA INTO TF DATASETS
minibatchsize = 32
train_x = train_pds.text  # used by keras tokenizers later
train_ds = tf.data.Dataset.from_tensor_slices((train_pds.text, train_pds.label))
train_ds = train_ds.batch(minibatchsize)  # needed to allow iteration
eval_ds = tf.data.Dataset.from_tensor_slices((eval_pds.text, eval_pds.label))  
eval_ds = eval_ds.batch(minibatchsize) 

# Test of some basic metrics to make sure our trainingset is in order.
print("*** balance:")  
print(train_pds.groupby(['source', 'label']).count())

print("\n*** check training sample in validation: ")
for i, s in enumerate(eval_pds.text):
    if len(train_pds[train_pds.text==s]):
        print('repetitive val. sample:', i)  
        print(s, "\n")

print("\n*** basic complexity:")
import nltk
nltk.download('punkt')
from nltk.tokenize import RegexpTokenizer, sent_tokenize
tkz = RegexpTokenizer(r'\w+')
train_pds["words"] = train_pds['text'].apply(lambda x: len(tkz.tokenize(x)))
train_pds["wordlen"] = train_pds['text'].str.len() / train_pds["words"]
train_pds["sentlen"] = train_pds["words"] / train_pds['text'].apply(lambda x: len(sent_tokenize(x)))

train_pds.groupby(['source', 'label']).mean()

# preprocessing / tokenizer for BoW model
max_tokens = 20000
text_vectorization = TextVectorization(
    max_tokens=max_tokens,
    output_mode="multi_hot",
)
text_vectorization.adapt(train_x)

train_v = train_ds.map(lambda x, y: (text_vectorization(x), y), num_parallel_calls=4)
eval_v = eval_ds.map(lambda x, y: (text_vectorization(x), y), num_parallel_calls=4)

vocab = text_vectorization.get_vocabulary()
print(len(vocab))  # we have 20k words
print(vocab[:10])  # words in order of commonality 
print(vocab[-10:]) 

# 2-layers fully connected architecture for BoW

def get_model(max_tokens=max_tokens, hidden_dim=16):
    inputs = keras.Input(shape=(max_tokens,))
    x = layers.Dense(hidden_dim, activation="relu")(inputs)
    x = layers.Dropout(0.5)(x)
    outputs = layers.Dense(1, activation="sigmoid")(x)
    model = keras.Model(inputs, outputs)
    model.compile(optimizer="rmsprop",
                  loss="binary_crossentropy",
                  metrics=["accuracy"])
    return model

model = get_model()
model.summary()

# Train model

model.fit(
    train_v,
    validation_data=eval_v,
    epochs=10,
    callbacks=[ModelCheckpoint('modelcheckpoint', save_best_only=True)])


# Test model. Note accuracy is .99
model = keras.models.load_model('modelcheckpoint')
print(model.evaluate(eval_v))

# SAVE MODEL
model.save("mbow-alldata")

pickle.dump({'config': text_vectorization.get_config(), 'weights': text_vectorization.get_weights()},
            open("textvectorizer.pickle", "wb"))

**Additional notes**

- We experimented with more complex models as well, but the accuracy of the model in this notebook was sufficient (.99 on test set)
- The model was tested on LS pages from states: 286 (78%) out of 365 pages received a prediction of >0.7
